Finalizing tests.

+++ contracts/MasterRegistryV2.sol
@@ -8,16 +8,16 @@ import '@openzeppelin/contracts-upgradeable/proxy/Initializable.sol';
  */
 contract MasterRegistryV2 is Initializable, OwnableUpgradeable {
 
-    event RecordChanged(address indexed series, uint8 indexed key, address value);
-    event ContentChanged(address indexed series, uint8 indexed key, string value);
+    event RecordChanged(address indexed series, uint16 indexed key, address value);
+    event ContentChanged(address indexed series, uint16 indexed key, string value);
 
     // Mapping PluginID => Pluggin contract address
-    mapping(uint8=>address) private plugins;
+    mapping(uint16=>address) private plugins;
     // Mapping Series Address => PluginID => Deployd Contract Address 
-    mapping(address=>mapping(uint8=>address)) private records;
+    mapping(address=>mapping(uint16=>address)) private records;
     // Mapping Series Address => PluginID => Content
-    mapping(address=>mapping(uint8=>string)) private contents;
-    uint8 public testVariable;
+    mapping(address=>mapping(uint16=>string)) private contents;
+    uint16 public testVariable;
 
     /**
      * Modifier that only allow the following entities change content:
@@ -27,7 +27,7 @@ contract MasterRegistryV2 is Initializable, OwnableUpgradeable {
      * @param _series The plugin index to update.
      * @param _key The new address where remains the plugin.
      */
-    modifier authorizedRecord(address _series, uint8 _key) {
+    modifier authorizedRecord(address _series, uint16 _key) {
         require(isSeriesOwner(_series) ||
         isRecordItself(_series, _key) || 
         isRecordPlugin(_series, _key), "Not authorized");
@@ -39,7 +39,7 @@ contract MasterRegistryV2 is Initializable, OwnableUpgradeable {
      * @param _series The plugin index to update.
      * @param _key The new address where remains the plugin.
      */
-    modifier onlySeriesOwner(address _series, uint8 _key) {
+    modifier onlySeriesOwner(address _series, uint16 _key) {
         require(isSeriesOwner(_series), "Not authorized");
         _;
     }
@@ -55,7 +55,7 @@ contract MasterRegistryV2 is Initializable, OwnableUpgradeable {
      * @param key The key to set.
      * @param value The text data value to set.
      */
-    function setRecord(address series, uint8 key, address value) public authorizedRecord(series, key) {
+    function setRecord(address series, uint16 key, address value) public authorizedRecord(series, key) {
         records[series][key] = value;
         emit RecordChanged(series, key, value);
     }
@@ -66,7 +66,7 @@ contract MasterRegistryV2 is Initializable, OwnableUpgradeable {
      * @param key The text data key to query.
      * @return The associated text data.
      */
-    function getRecord(address series, uint8 key) public view returns (address) {
+    function getRecord(address series, uint16 key) public view returns (address) {
         return records[series][key];
     }
 
@@ -77,7 +77,7 @@ contract MasterRegistryV2 is Initializable, OwnableUpgradeable {
      * @param key The key to set.
      * @param value The text data value to set.
      */
-    function setContent(address series, uint8 key, string memory value) public onlySeriesOwner(series, key) {
+    function setContent(address series, uint16 key, string memory value) public onlySeriesOwner(series, key) {
         contents[series][key] = value;
         emit ContentChanged(series, key, value);
     }
@@ -88,7 +88,7 @@ contract MasterRegistryV2 is Initializable, OwnableUpgradeable {
      * @param key The text data key to query.
      * @return The associated text data.
      */
-    function getContent(address series, uint8 key) public view returns (string memory) {
+    function getContent(address series, uint16 key) public view returns (string memory) {
         return contents[series][key];
     }
 
@@ -98,7 +98,7 @@ contract MasterRegistryV2 is Initializable, OwnableUpgradeable {
      * @param pluginID The plugin index to update.
      * @param pluginAddress The new address where remains the plugin.
      */
-    function setPluginController(uint8 pluginID, address pluginAddress) public onlyOwner {
+    function setPluginController(uint16 pluginID, address pluginAddress) public onlyOwner {
         plugins[pluginID] = pluginAddress;
     }
 
@@ -106,11 +106,11 @@ contract MasterRegistryV2 is Initializable, OwnableUpgradeable {
         return OwnableUpgradeable(_series).owner() == _msgSender();
     }
 
-    function isRecordItself(address _series, uint8 _key) private view returns (bool) {
+    function isRecordItself(address _series, uint16 _key) private view returns (bool) {
         return records[_series][_key] == _msgSender();
     }
 
-    function isRecordPlugin(address _series, uint8 _key) private view returns (bool) {
+    function isRecordPlugin(address _series, uint16 _key) private view returns (bool) {
         return _msgSender() == plugins[_key] && records[_series][_key] == address(0);
     }
 }
