+++ contracts/utils/IMasterRegistry.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.6.0;
 
 interface IMasterRegistry {
-    function setRecord(address series, uint8 key, address value) external;
+    function setRecord(address series, uint16 key, address value) external;
 }
